The big rename (calendar-picker -> times-square)
