+++ CONTRIBUTING.md
@@ -1,7 +1,7 @@
 Contributing
 ============
 
-If you would like to contribute code to CalendarPicker you can do so through GitHub by
+If you would like to contribute code to TimesSquare you can do so through GitHub by
 forking the repository and sending a pull request.
 
 When submitting code, please make every effort to follow existing conventions
