+++ library/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.squareup.calendar"
+    package="com.squareup.timessquare"
     android:versionCode="1"
     android:versionName="1.0">
