+++ library/res/layout/month.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 
-<com.squareup.calendar.MonthView
+<com.squareup.timessquare.MonthView
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:orientation="vertical"
     android:layout_width="fill_parent"
@@ -15,12 +15,12 @@
       android:layout_marginBottom="@dimen/calendar_month_title_bottommargin"
       style="@style/CalendarTitle"
       />
-  <com.squareup.calendar.CalendarGridView
+  <com.squareup.timessquare.CalendarGridView
       android:id="@+id/calendar_grid"
       android:layout_width="match_parent"
       android:layout_height="wrap_content"
       >
-    <com.squareup.calendar.CalendarRowView
+    <com.squareup.timessquare.CalendarRowView
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:paddingBottom="@dimen/calendar_day_headers_paddingbottom"
@@ -60,12 +60,12 @@
           android:layout_height="wrap_content"
           style="@style/CalendarCell.DayHeader"
           />
-    </com.squareup.calendar.CalendarRowView>
+    </com.squareup.timessquare.CalendarRowView>
     <include layout="@layout/week"/>
     <include layout="@layout/week"/>
     <include layout="@layout/week"/>
     <include layout="@layout/week"/>
     <include layout="@layout/week"/>
     <include layout="@layout/week"/>
-  </com.squareup.calendar.CalendarGridView>
-</com.squareup.calendar.MonthView>
+  </com.squareup.timessquare.CalendarGridView>
+</com.squareup.timessquare.MonthView>
