+++ library/res/layout/week.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 
-<com.squareup.calendar.CalendarRowView
+<com.squareup.timessquare.CalendarRowView
     xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
@@ -40,4 +40,4 @@
       android:layout_height="wrap_content"
       style="@style/CalendarCell.CalendarDate"
       />
-</com.squareup.calendar.CalendarRowView>
+</com.squareup.timessquare.CalendarRowView>
