+++ sample/pom.xml
@@ -7,13 +7,13 @@
 
   <parent>
     <groupId>com.squareup</groupId>
-    <artifactId>calendar-parent</artifactId>
+    <artifactId>times-square-parent</artifactId>
     <version>1.0-SNAPSHOT</version>
     <relativePath>../pom.xml</relativePath>
   </parent>
 
-  <artifactId>calendar-sample</artifactId>
-  <name>Android-CalendarPicker Sample</name>
+  <artifactId>timessquare-sample</artifactId>
+  <name>Android-TimesSquare Sample</name>
   <packaging>apk</packaging>
 
   <dependencies>
@@ -25,7 +25,7 @@
 
     <dependency>
       <groupId>com.squareup</groupId>
-      <artifactId>android-calendar-picker</artifactId>
+      <artifactId>android-times-square</artifactId>
       <version>${project.version}</version>
       <type>apklib</type>
     </dependency>
