+++ sample/res/layout/calendar_picker.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     >
-  <com.squareup.calendar.CalendarPickerView
+  <com.squareup.timessquare.CalendarPickerView
       android:id="@+id/calendar_view"
       android:layout_width="match_parent"
       android:layout_height="0dp"
