+++ sample/src/com/squareup/timessquare/sample/SampleTimesSquareActivity.java
@@ -1,4 +1,4 @@
-package com.squareup.calendar.sample;
+package com.squareup.timessquare.sample;
 
 import android.app.Activity;
 import android.os.Bundle;
@@ -6,14 +6,14 @@ import android.util.Log;
 import android.view.View;
 import android.view.View.OnClickListener;
 import android.widget.Toast;
-import com.squareup.calendar.CalendarPickerView;
+import com.squareup.timessquare.CalendarPickerView;
 import java.util.Calendar;
 import java.util.Date;
 
 import static android.widget.Toast.LENGTH_SHORT;
 
-public class SampleCalendarPickerActivity extends Activity {
-  private static final String TAG = "SampleCalendarPickerActivity";
+public class SampleTimesSquareActivity extends Activity {
+  private static final String TAG = "SampleTimesSquareActivity";
 
   @Override protected void onCreate(Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
@@ -29,7 +29,7 @@ public class SampleCalendarPickerActivity extends Activity {
       @Override public void onClick(View view) {
         Log.d(TAG, "Selected time in millis: " + calendar.getSelectedDate().getTime());
         String toast = "Selected: " + calendar.getSelectedDate().getTime();
-        Toast.makeText(SampleCalendarPickerActivity.this, toast, LENGTH_SHORT).show();
+        Toast.makeText(SampleTimesSquareActivity.this, toast, LENGTH_SHORT).show();
       }
     });
   }
